requirements.txt
# Required dependencies for the project
ray
duckdb
pytest
pytest-mock
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-prometheus
logging
